refactor: rename src/workflow/ to src/agent/ (closes #237)
- Move src/workflow/ → src/agent/
- Update all internal imports (from workflow.* → from agent.*)
- Update patch targets in tests (workflow.executor.* → agent.executor.*)
- Update pyproject.toml: packages list and plan-execute entry point
- Update INSTRUCTIONS.md and CLAUDE.md references
