+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/workflow", "src/llm"]
+packages = ["src/agent", "src/llm"]
 
 [project]
 name = "assetopsbench-mcp"
@@ -27,7 +27,7 @@ dependencies = [
 ]
 
 [project.scripts]
-plan-execute = "workflow.cli:main"
+plan-execute = "agent.cli:main"
 iot-mcp-server = "servers.iot.main:main"
 utilities-mcp-server = "servers.utilities.main:main"
 fmsr-mcp-server = "servers.fmsr.main:main"
